+++ packages/user-operation-controller/src/utils/validation.ts
@@ -135,11 +135,12 @@ export function validatePrepareUserOperationResponse(
 export function validateUpdateUserOperationResponse(
   response: UpdateUserOperationResponse,
 ) {
-  const HexOrEmptyBytes = defineHex();
-
   const ValidResponse = optional(
     object({
-      paymasterAndData: optional(HexOrEmptyBytes),
+      paymasterAndData: optional(defineHexOrEmptyBytes()),
+      callGasLimit: optional(defineHexOrEmptyBytes()),
+      preVerificationGas: optional(defineHexOrEmptyBytes()),
+      verificationGasLimit: optional(defineHexOrEmptyBytes()),
     }),
   );